+++ StabilityMatrix.Avalonia/DesignData/DesignData.cs
@@ -1098,6 +1098,19 @@ The gallery images are often inpainted, but you will get something very similar 
             }
         };
 
+    public static ModelMetadataEditorDialogViewModel MetadataEditorDialogViewModel =>
+        DialogFactory.Get<ModelMetadataEditorDialogViewModel>(vm =>
+        {
+            vm.ThumbnailFilePath = Assets.NoImage.ToString();
+            vm.Tags = "tag1, tag2, tag3";
+            vm.ModelDescription = "This is a description";
+            vm.ModelName = "Model Name";
+            vm.VersionName = "1.0.0";
+            vm.TrainedWords = "word1, word2, word3";
+            vm.ModelType = CivitModelType.Checkpoint;
+            vm.BaseModelType = CivitBaseModelType.Pony;
+        });
+
     public static string CurrentDirectory => Directory.GetCurrentDirectory();
 
     public static Indexer Types { get; } = new();